Move footer component to layout component
Fix laptop image in top categories
Add more new phone images
Change padding and margins at Details Comp.
