+++ src/components/core/Layout.jsx
@@ -1,11 +1,13 @@
 import PropTypes from 'prop-types'
 import Header from './Header';
+import Footer from './Footer';
 
 const Layout = ({ renderHeader, children }) => {
     return (
         <div className="md:bg-gray-50">
             {renderHeader && <Header />}
             {children}
+            <Footer />
         </div>
     );
 }
